requirements.txt
psycopg2-binary
pytest
pytest-html
python-dotenv
